move token-relevant utilities to http-json-testing lib
